Add files via upload
The add-to-cart button is a feature of ecommerce stores that 
allows customers to choose items to purchase without actually 
completing the payment. For online stores, it lives on individual 
product pages, functioning as the digital equivalent of a shopping 
cart in a brick and mortar store. the add-to-cart button allows 
customers to save items to their cart so they can continue 
shopping, then complete the checkout process later on in the 
process, resulting in a more seamless customer experience. It 
also allows shoppers to purchase multiple items at one time.